⚡ Bolt: Centralize motion preference hook and stabilize task callbacks
- Created src/hooks/usePrefersReducedMotion.ts using useSyncExternalStore.
- Refactored Button, ScrollToTop, Tooltip, LoadingSpinner, Skeleton, BlueprintDisplay, and ToastContainer to use the centralized hook.
- Stabilized handleToggleTaskStatus in useTaskManagement.ts by using dataRef.current to prevent O(N) re-renders when toggling tasks.
- Improved code maintainability and interaction performance.
