+++ src/hooks/useTaskManagement.ts
@@ -201,12 +201,16 @@ export function useTaskManagement(ideaId: string): UseTaskManagementReturn {
       const newStatus: TaskStatus =
         currentStatus === 'completed' ? 'todo' : 'completed';
 
+      // PERFORMANCE: Use dataRef.current instead of data from outer scope
+      // to keep this callback identity stable across re-renders.
+      const currentData = dataRef.current;
+
       // Store previous state for potential rollback
-      previousDataRef.current = data;
+      previousDataRef.current = currentData;
 
       // Find the task for toast message BEFORE making changes
       const findTask = () => {
-        return data?.deliverables
+        return currentData?.deliverables
           .flatMap((d) => d.tasks)
           .find((t) => t.id === taskId);
       };
@@ -293,7 +297,7 @@ export function useTaskManagement(ideaId: string): UseTaskManagementReturn {
         setUpdatingTaskId(null);
       }
     },
-    [data, logger, applyTaskStatusUpdate]
+    [logger, applyTaskStatusUpdate]
   );
 
   // Toggle deliverable expansion
